Changes to backend and front end
The swipe is a total fucking pain in the ass and the library we are
using is broken partially. Spent a while trying to get it working. App
still needs a lot of work as the shortlist, what is the go with this?

We also need to fix the data in the backend as if a venue doesn't have a
location associated, it won't show in the search results.