Changes to backend and huge re-write of internals for ionic app.
Backend now provides an updated JSON output to the client to  allow the
client to calculate distance from a venue. Also massive re-write of
controller and services code. Fixes to various bugs which cause
stateless rubbish. Introduction to client CacheFactory which assists in
data caching of queries from backend server as well as singular source
for controllers.